Skilt ut SketchArea i en separat component
